requirements.txt
django==3.0.5
Celery==4.4.2
redis==3.5.0
requests==2.23.0
beautifulsoup4==4.9.0
lxml==4.5.0
djangorestframework==3.11.0
django-filter==2.2.0
